requirements.txt
pydantic-ai
gradio >= 5.9.0
python-dotenv
openai